- Fixed segfault (due to capital S in format string instead of lowercase s) - Added format string checking for log functions - Fixed all format string errors
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5905 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ reader-cryptoworks.c
@@ -221,7 +221,7 @@ static int32_t cryptoworks_card_init(struct s_reader * reader, ATR newatr)
   if (read_record(reader, 0x80, cta_res)>=7)		// read serial
     memcpy(reader->hexserial, cta_res+2, 5);
   cs_ri_log (reader, "type: CryptoWorks, caid: %04X, ascii serial: %llu, hex serial: %s",
-            reader->caid, b2ll(5, reader->hexserial),cs_hexdump(0, reader->hexserial, 5, tmp, sizeof(tmp)));
+            reader->caid, (unsigned long long) b2ll(5, reader->hexserial),cs_hexdump(0, reader->hexserial, 5, tmp, sizeof(tmp)));
 
   if (read_record(reader, 0x9E, cta_res)>=66)	// read ISK
   {
